requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
infobip-api-python-client==3.0.3
marshmallow==3.21.1
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
